use static cast to suppress type cast warning #82

+++ tiny_cnn/layers/fully_connected_layer.h
@@ -75,7 +75,7 @@ public:
     }
 
     const vec_t& back_propagation(const vec_t& curr_delta, size_t index) override {
-        const vec_t& prev_out = prev_->output(index);
+        const vec_t& prev_out = prev_->output(static_cast<int>(index));
         const activation::function& prev_h = prev_->activation_function();
         vec_t& prev_delta = prev_delta_[index];
         vec_t& dW = dW_[index];
